# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowCyt Panel API",
    "description": "Hybrid AI tool for multi-color flow cytometry panel design.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/quality-registry/issues": {
      "post": {
        "tags": [
          "quality-registry"
        ],
        "summary": "Create Issue",
        "description": "Create a new quality issue and trigger projection update.",
        "operationId": "create_issue_api_v1_quality_registry_issues_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityIssueCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/panel-history": {
      "get": {
        "tags": [
          "admin",
          "admin-panel-history"
        ],
        "summary": "List History",
        "operationId": "list_history_api_v1_admin_panel_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelHistoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/panel-history/{entry_id}": {
      "get": {
        "tags": [
          "admin",
          "admin-panel-history"
        ],
        "summary": "Get History Detail",
        "operationId": "get_history_detail_api_v1_admin_panel_history__entry_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelHistoryDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/quality-registry/issues/{issue_id}": {
      "put": {
        "tags": [
          "admin",
          "admin-quality-registry"
        ],
        "summary": "Update Issue",
        "operationId": "update_issue_api_v1_admin_quality_registry_issues__issue_id__put",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityIssueUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin",
          "admin-quality-registry"
        ],
        "summary": "Get Issue",
        "description": "Get a single quality issue by ID.",
        "operationId": "get_issue_api_v1_admin_quality_registry_issues__issue_id__get",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/quality-registry/issues": {
      "get": {
        "tags": [
          "admin",
          "admin-quality-registry"
        ],
        "summary": "List Issues",
        "description": "List quality issues, optionally filtered by status.",
        "operationId": "list_issues_api_v1_admin_quality_registry_issues_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityIssueResponse"
                  },
                  "title": "Response List Issues Api V1 Admin Quality Registry Issues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/quality-registry/issues/{issue_id}/history": {
      "get": {
        "tags": [
          "admin",
          "admin-quality-registry"
        ],
        "summary": "Get History",
        "description": "Get audit history for a quality issue.",
        "operationId": "get_history_api_v1_admin_quality_registry_issues__issue_id__history_get",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "title": "Response Get History Api V1 Admin Quality Registry Issues  Issue Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/quality-registry/review-queue/{issue_id}/resolve": {
      "post": {
        "tags": [
          "admin",
          "admin-quality-registry"
        ],
        "summary": "Resolve Review",
        "description": "Resolve a manual review item.",
        "operationId": "resolve_review_api_v1_admin_quality_registry_review_queue__issue_id__resolve_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuditEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "issue_id": {
            "type": "string",
            "title": "Issue Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "issue_id",
          "action",
          "actor"
        ],
        "title": "AuditEvent",
        "description": "Immutable record of a single mutation on a quality issue."
      },
...
      "EntityKey": {
        "properties": {
          "species": {
            "type": "string",
            "title": "Species"
          },
          "normalized_marker": {
            "type": "string",
            "title": "Normalized Marker"
          },
          "clone": {
            "type": "string",
            "title": "Clone"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "catalog_number": {
            "type": "string",
            "title": "Catalog Number"
          },
          "lot_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Number"
          }
        },
        "type": "object",
        "required": [
          "species",
          "normalized_marker",
          "clone",
          "brand",
          "catalog_number"
        ],
        "title": "EntityKey",
        "description": "Canonical identity for a specific antibody reagent.\n\nEquality & hashing are based on the 5 required fields.\n``lot_number`` is tracked as metadata but is NOT part of identity."
      },
      "FeedbackKey": {
        "properties": {
          "species": {
            "type": "string",
            "title": "Species"
          },
          "normalized_marker": {
            "type": "string",
            "title": "Normalized Marker"
          },
          "fluorochrome": {
            "type": "string",
            "title": "Fluorochrome"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "clone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clone"
          }
        },
        "type": "object",
        "required": [
          "species",
          "normalized_marker",
          "fluorochrome",
          "brand"
        ],
        "title": "FeedbackKey",
        "description": "Identity key for a quality-feedback submission.\n\nEquality & hashing are based on all non-None fields, so two submissions\nthat differ only in clone (where one has clone=None) are *not* equal."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PanelHistoryDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "species": {
            "type": "string",
            "title": "Species"
          },
          "inventory_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventory File"
          },
          "requested_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requested Markers"
          },
          "missing_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Markers"
          },
          "selected_panel": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Selected Panel"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "api_base": {
            "type": "string",
            "title": "Api Base"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "species",
          "requested_markers",
          "missing_markers",
          "selected_panel",
          "rationale",
          "model_name",
          "api_base"
        ],
        "title": "PanelHistoryDetail"
      },
      "PanelHistoryDetailResponse": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/PanelHistoryDetail"
          }
        },
        "type": "object",
        "required": [
          "item"
        ],
        "title": "PanelHistoryDetailResponse"
      },
      "PanelHistoryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PanelHistorySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "PanelHistoryListResponse"
      },
      "PanelHistorySummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "species": {
            "type": "string",
            "title": "Species"
          },
          "inventory_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventory File"
          },
          "requested_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requested Markers"
          },
          "missing_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Markers"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "species",
          "requested_markers",
          "missing_markers",
          "model_name"
        ],
        "title": "PanelHistorySummary"
      },
      "QualityIssueCreate": {
        "properties": {
          "issue_text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Issue Text"
          },
          "reported_by": {
            "type": "string",
            "minLength": 1,
            "title": "Reported By"
          },
          "species": {
            "type": "string",
            "minLength": 1,
            "title": "Species"
          },
          "marker": {
            "type": "string",
            "minLength": 1,
            "title": "Marker"
          },
          "fluorochrome": {
            "type": "string",
            "minLength": 1,
            "title": "Fluorochrome"
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand"
          },
          "clone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clone"
          }
        },
        "type": "object",
        "required": [
          "issue_text",
          "reported_by",
          "species",
          "marker",
          "fluorochrome",
          "brand"
        ],
        "title": "QualityIssueCreate",
        "description": "Payload for submitting a new quality issue."
      },
      "QualityIssueResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "feedback_key": {
            "$ref": "#/components/schemas/FeedbackKey"
          },
          "entity_key": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityKey"
              },
              {
                "type": "null"
              }
            ]
          },
          "issue_text": {
            "type": "string",
            "title": "Issue Text"
          },
          "reported_by": {
            "type": "string",
            "title": "Reported By"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "submitted"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "feedback_key",
          "issue_text",
          "reported_by",
          "created_at",
          "updated_at"
        ],
        "title": "QualityIssueResponse",
        "description": "Full representation of a persisted quality issue."
      },
      "QualityIssueUpdate": {
        "properties": {
          "issue_text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Issue Text"
          },
          "reported_by": {
            "type": "string",
            "minLength": 1,
            "title": "Reported By"
          }
        },
        "type": "object",
        "required": [
          "issue_text",
          "reported_by"
        ],
        "title": "QualityIssueUpdate",
        "description": "Payload for editing an existing quality issue.\n\nOnly ``issue_text`` and ``reported_by`` may be changed.\n``feedback_key``, ``entity_key``, and ``status`` are immutable via this schema."
      },
      "ResolveReviewRequest": {
        "properties": {
          "reviewer": {
            "type": "string",
            "minLength": 1,
            "title": "Reviewer"
          },
          "entity_key": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityKey"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "reviewer"
        ],
        "title": "ResolveReviewRequest",
        "description": "Resolve a manual review item."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}